refactor(three): consolidate Scene hooks under one module (closes #57)
Move the nine three.js / R3F hooks that the World viewport (and its
TrafficData / TriggerData / StreetData / AI Sections / PolygonSoup /
RotationVisualizer siblings) depend on out of the alphabetically-mixed
src/hooks/ folder into a cohesive src/lib/three/scene/ directory:

- useApplyDisplayMatrix
- useApplyMatrixToObject
- useAutoFitCamera
- useCameraBridgeSync
- useFitMapBounds
- useFlyCameraToTarget
- useLineSegmentsGeometry
- useSceneEnvironment
- useUpdateInstancedMesh

Pure move — hook signatures and implementations unchanged. git mv
preserves per-file history. Adds a src/lib/three/scene/index.ts barrel
re-export; callers continue to import from the specific hook file so
the leaflet-backed useFitMapBounds doesn't drag leaflet (which touches
window at import time and breaks node-side tests) into siblings that
don't need it.

Updated 12 caller import sites and one cross-reference comment in
SelectionOutline.tsx. useInstancedSelection still uses
useUpdateInstancedMesh as its underlying primitive.

Test counts identical to baseline (1089 passing / 9 unrelated ENOENT
failures from missing example/older builds/ fixtures). TypeScript
errors identical (39, all in pre-existing TrafficDataViewport /
SectionsTab / trafficDataExtensions / material.test / wheelGraphicsSpec.test).
