+++ src/lib/three/scene/useApplyMatrixToObject.ts
