+++ src/lib/three/scene/useAutoFitCamera.ts
