+++ src/lib/three/scene/useCameraBridgeSync.ts
