+++ src/lib/three/scene/useFitMapBounds.ts
