+++ src/lib/three/scene/useLineSegmentsGeometry.ts
