+++ src/lib/three/scene/useSceneEnvironment.ts
