+++ src/lib/three/scene/useUpdateInstancedMesh.ts
